ai-apps/2025-yearend-app/llm-utility-app/llm-utility-api/pyproject.toml
[project]
name = "llm-utility-api"
version = "0.1.0"
description = "A minimal FastAPI-based LLM utility API (health, summarize, rewrite, extract)."
authors = [
  { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
  "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]



